Aranging namespaces and structure of new BasicCompletes implementation.
