+++ src/Vlingo.Common/ICompletes.cs
@@ -6,6 +6,7 @@
 // one at https://mozilla.org/MPL/2.0/.
 
 using System;
+using Vlingo.Common.Completion;
 
 namespace Vlingo.Common
 {
