+++ src/Vlingo.Common/Vlingo.Common.csproj
@@ -29,8 +29,5 @@
     <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
     <PackageReference Include="Vlingo.UUID" Version="1.0.0" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Completion" />
-  </ItemGroup>
 
 </Project>
